requirements.txt
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-oauth-toolkit==1.7.1
twilio==8.10.3
africastalking==1.2.5
celery==5.3.4
redis==5.0.1
python-decouple==3.8
Pillow==10.1.0
psutil==5.9.6
requests==2.31.0